Move UrlValidator to core.
