+++ api/app/models/spree/webhooks/endpoint.rb
@@ -1,7 +1,7 @@
 module Spree
   module Webhooks
     class Endpoint < Spree::Webhooks::Base
-      validates :url, :'spree/webhooks/validators/url' => true
+      validates :url, url: true
     end
   end
 end
